refactor(schemas): move schema ownership out of the root entrypoint
lib/hive.rb co-located the entire Hive::Schemas namespace
(SCHEMA_VERSIONS, ErrorEnvelope, EnvelopeEmitter, closed enums) with
root-entrypoint concerns, making every published-contract version bump
a change to the root file. Move the namespace verbatim into its own
lib/hive/schemas.rb and require_relative it from the root entrypoint,
keeping the public constant path unchanged for every require "hive"
consumer. Only schema_dir's relative depth changes to match its new
location.

Add regression tests asserting the dedicated file owns the namespace,
that requiring hive loads it exactly once from its own file, and that
schema_dir/schema_path still resolve to the published schemas.
